create new test file
